requirements.txt
easydict>=1.9
ipython>=7.8.0
matplotlib>=3.1.1
numpy>=1.16.1
opencv-python>=4.1.1.26
Pillow>=6.1.0
scikit-image>=0.17.2
scikit-learn>=0.22.1
scipy>=1.3.1

torchdiffeq>=0.0.1

tqdm>=4.19.9